+++ x/cwfees/types/keys.go
@@ -2,7 +2,7 @@ package types
 
 import "cosmossdk.io/collections"
 
-const ModuleName = "cwgrant"
+const ModuleName = "cwfees"
 
 var (
 	GrantersPrefix = collections.NewPrefix(0)